Updated Russian translations for v.2.8.0
* Updated strings in nvda.po.
* Updated translation of readme.md.

Note:
After updating the localization, it is not necessary to release a full update with changes to the changelog text in \`manifest.ini\`. It is possible to simply release versions such as 2.8.1, if translators send you more updated translations, then release 2.8.2, 2.8.3, etc.
Sending out updates for updated localizations does not need to be reflected in the version journal.
This way, there will be no constant delay in translations specifically for these moments.